--- packages/certifier/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "certifier"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-doctest = false
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bench]]
-name = "certifier"
-harness = false
-
-[dependencies]
-ahash = "0.8.2"
-
-async-trait = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-# Logging
-log = { workspace = true }
-env_logger = { workspace = true }
-
-# Error
-thiserror = "1.0.31"
-
-# Test
-mockall = "0.11.0"
-criterion = "0.4"
-
-
-[dev-dependencies]
-tokio-test = "0.4"
