--- packages/certifier/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod certifier;
-pub mod certifier_candidate;
-pub mod core;
-
-// Unit Tests
-#[cfg(test)]
-mod tests;
