+++ packages/talos_core/Cargo.toml
@@ -7,6 +7,9 @@ edition = "2021"
 #
 strum = { version = "0.24", features = ["derive"] }
 
+#  Ahash hashmap
+ahash = "0.8.3"
+
 # Logging
 log = { workspace = true }
 env_logger = { workspace = true }
@@ -23,6 +26,10 @@ serde_json = { workspace = true }
 thiserror = "1.0.31"
 
 # internal crates
-certifier = { path = "../certifier" }
 suffix = { path = "../suffix" }
 logger = { path = "../logger" }
+
+
+[dev-dependencies]
+tokio-test = "0.4"
+criterion = "0.4"
