+++ packages/talos_core/benches/certifier.rs
@@ -1,5 +1,5 @@
-use certifier::{certifier::Certifier, certifier_candidate::CertifierCandidate, core::convert_vec_to_hashmap};
 use criterion::{criterion_group, criterion_main, Criterion};
+use talos_core::{certifier::utils::convert_vec_to_hashmap, Certifier, CertifierCandidate};
 
 fn bench_certifier_rule_1(c: &mut Criterion) {
     c.bench_function("[CERTIFIER - Rule 1] Unconditional Commit", |b| {
