+++ packages/talos_core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod certifier;
 pub mod config;
 pub mod core;
 pub mod errors;
@@ -5,3 +6,5 @@ pub mod model;
 pub mod ports;
 
 pub use crate::core::{ChannelMessage, SystemMessage};
+pub use certifier::Certifier;
+pub use certifier::CertifierCandidate;
